+++ CHANGELOG.md
@@ -29,3 +29,4 @@ with `Angel.eagerParseRequestBodies`.
 replaced with `parseBody`, `parseUploadedFiles`, and `parseRawRequestBuffer`, respectively.
 * Removed the synchronous equivalents of the above methods (`body`, `files`, and `originalBuffer`),
 as well as `query`.
+* Remove `Angel.injections` and `RequestContext.injections`.